# Proyecto Final/CustomerService/customer-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Customer - OpenAPI 3.0
  description: |-
    <div style="text-align: justify;">
      <h3>Maestría en Software - Diseño de patrones de API</h3>
    
      <img src="https://blogposgrados.ups.edu.ec/hs-fs/hubfs/Logo-Posgrados.png?width=1517&height=262&name=Logo-Posgrados.png" alt="Image" align="right" width="400"/>
    
      <p>This is a contract to manage customer data based on the OpenAPI 3.0 specification.This is a contract that uses the API First-Contract principle to standardize the creation of endpoints to be used in our service and improve the development of our APIs.</p> 
    </div>
  termsOfService: http://swagger.io/terms/
  contact:
    email: ntellon@est.ups.edu.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0 - SNAPSHOT

externalDocs:
  description: Historia de usuario
  url: https://github.com/users/bvegaM/projects/11/views/1?pane=issue&itemId=79803180

servers:
  - url: http://localhost:8084/api/v1
    description: service
  - url: http://localhost:8081/api/v1
    description: customer-middleend
  - url: http://localhost:8080/api/v1
    description: api-gateway

tags:
  - name: customer
    description: Everything about customers

paths:
  /customers:
    post:
      tags:
        - customer
      summary: Create a new customer
      description: Create a new customer with an Unique ID
      operationId: createCustomer
      parameters:
        - name: x-cm-client-request-id
          in: header
          description: Unique ID for the request
          required: true
          schema:
            type: string
        - name: x-cm-client-user-agent
          in: header
          description: Type of device that consume the endpoint
          required: true
          schema:
            type: string
      requestBody:
        description: Created customer object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDto'
      responses:
        '201':
          description: succesfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '400':
          description: The server could not understand the request because of incorrect syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /customers/{id}:
    get:
      tags:
        - customer
      summary: Get an customer by ID
      description: Get a customer with an Unique ID
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          description: Unique ID of the account
          required: true
          schema:
            type: integer
            format: int64
        - name: x-cm-client-request-id
          in: header
          description: Unique ID for the request
          required: true
          schema:
            type: string
        - name: x-cm-client-user-agent
          in: header
          description: Type of device that consume the endpoint
          required: true
          schema:
            type: string
      responses:
        '201':
          description: succesfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: customer does not exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: The server could not understand the request because of incorrect syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /customers/identification/{identificacion}:
    get:
      tags:
        - customer
      summary: Get an customer by ID
      description: Get a customer with an Unique ID
      operationId: getCustomerByIdentification
      parameters:
        - name: identificacion
          in: path
          description: Unique IDENTIFICATION of the account
          required: true
          schema:
            type: string
            maxLength: 13
            pattern: "^[0-9]+$"
        - name: x-cm-client-request-id
          in: header
          description: Unique ID for the request
          required: true
          schema:
            type: string
        - name: x-cm-client-user-agent
          in: header
          description: Type of device that consume the endpoint
          required: true
          schema:
            type: string
      responses:
        '201':
          description: succesfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: customer does not exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: The server could not understand the request because of incorrect syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    CustomerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        identification:
          type: string
          maxLength: 13
          pattern: "^[0-9]+$"
          description: Identification must contain only numbers with a max length of 13 digits.
        firstName:
          type: string
          pattern: "^[a-zA-Z\\s]+$"
          description: First name must contain only letters and spaces.
        lastName:
          type: string
          pattern: "^[a-zA-Z\\s]+$"
          description: Last name must contain only letters and spaces.
        genre:
          type: string
          enum:
            - MALE
            - FEMALE
      required:
        - identification
        - firstName
        - lastName
        - genre
    Error:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        status:
          type: integer
    ErrorDetail:
      type: object
      properties:
        message:
          type: string